[REF] Bar de vie

+++ src/Client/App.h
@@ -43,7 +43,7 @@ public:
 	void ClearChatMessages();
 
 	void RenderOtherPlayersHealth();
-	std::string MakeHpBar(float currentHealth, float maxHealth);
+	std::string MakeHpBar(int currentHealth, int maxHealth);
 
 private:
 	std::vector<ChatLine> s_chatMessages;
